addDashboardPage
