+++ src/components/MobileTabl/MobileTab.jsx
@@ -13,16 +13,13 @@ const MobileTabl = ({ data }) => {
     data &&
     data.map(({ type, category, comment, sum, balance, date }) => {
       return (
-        <ContainerMobileTabl
-          sx={{ minWidth: 280 }}
-          isType={type === 'true' ? '+' : '-'}
-        >
+        <ContainerMobileTabl sx={{ minWidth: 280 }} isType={type ? '+' : '-'}>
           <CardBox key={unid}>
             <TypographyCard>
               Date <Span> {date}</Span>
             </TypographyCard>
             <TypographyCard>
-              Type <Span>{type === 'true' ? '+' : '-'}</Span>
+              Type <Span>{type ? '+' : '-'}</Span>
             </TypographyCard>
             <TypographyCard>
               Category <Span>{category}</Span>
@@ -31,7 +28,7 @@ const MobileTabl = ({ data }) => {
               Comment <Span>{comment}</Span>
             </TypographyCard>
             <TypographyCard>
-              Sum <Span isType={type === 'true' ? '+' : '-'}>{sum}</Span>
+              Sum <Span isType={type ? '+' : '-'}>{sum}</Span>
             </TypographyCard>
             <TypographyCard>
               Balance <Span>{balance}</Span>
